requirements.txt
dearpygui
python-obd
pyserial